+++ cmd/root.go
@@ -12,13 +12,17 @@ import (
 	"os"
 	"os/signal"
 	"path"
+	"path/filepath"
 	"strings"
 	"syscall"
 	"time"
 
+	"atomicgo.dev/keyboard"
+	"atomicgo.dev/keyboard/keys"
 	"connectrpc.com/connect"
-	tea "github.com/charmbracelet/bubbletea"
 	"github.com/getsentry/sentry-go"
+	"github.com/go-jose/go-jose/v4"
+	josejwt "github.com/go-jose/go-jose/v4/jwt"
 	"github.com/google/uuid"
 	"github.com/overmindtech/cli/tracing"
 	"github.com/overmindtech/sdp-go"
@@ -249,82 +253,6 @@ func Execute() {
 	}
 }
 
-type statusMsg int
-
-const (
-	PromptUser             statusMsg = 0
-	WaitingForConfirmation statusMsg = 1
-	Authenticated          statusMsg = 2
-	ErrorAuthenticating    statusMsg = 3
-)
-
-type authenticateModel struct {
-	ctx context.Context
-
-	status     statusMsg
-	err        error
-	deviceCode *oauth2.DeviceAuthResponse
-	config     oauth2.Config
-	token      *oauth2.Token
-
-	width int
-}
-
-func (m authenticateModel) Init() tea.Cmd {
-	return openBrowserCmd(m.deviceCode.VerificationURI)
-}
-
-func (m authenticateModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	cmds := []tea.Cmd{}
-
-	switch msg := msg.(type) {
-	case tea.WindowSizeMsg:
-		m.width = min(MAX_TERMINAL_WIDTH, msg.Width)
-
-	case tea.KeyMsg:
-		switch msg.String() {
-		default:
-			{
-				if m.status == Authenticated {
-					return m, tea.Quit
-				}
-			}
-		case "ctrl+c", "q":
-			return m, tea.Quit
-		}
-
-	case *oauth2.Token:
-		m.status = Authenticated
-		m.token = msg
-
-	case statusMsg:
-		switch msg {
-		case PromptUser:
-			cmds = append(cmds, openBrowserCmd(m.deviceCode.VerificationURI))
-		case WaitingForConfirmation:
-			m.status = WaitingForConfirmation
-			cmds = append(cmds, awaitToken(m.ctx, m.config, m.deviceCode))
-		case Authenticated:
-		case ErrorAuthenticating:
-		}
-
-	case displayAuthorizationInstructionsMsg:
-		m.status = WaitingForConfirmation
-		cmds = append(cmds, awaitToken(m.ctx, m.config, m.deviceCode))
-
-	case failedToAuthenticateErrorMsg:
-		m.err = msg.err
-		m.status = ErrorAuthenticating
-		cmds = append(cmds, tea.Quit)
-
-	case errMsg:
-		m.err = msg.err
-		cmds = append(cmds, tea.Quit)
-	}
-
-	return m, tea.Batch(cmds...)
-}
-
 const beginAuthMessage string = `# Authenticate with a browser
 
 Attempting to automatically open the SSO authorization page in your default browser.
@@ -336,47 +264,6 @@ Then enter the code:
 
 	%v
 `
-
-func (m authenticateModel) View() string {
-	var output string
-
-	switch m.status {
-	case PromptUser, WaitingForConfirmation:
-		prompt := fmt.Sprintf(beginAuthMessage, m.deviceCode.VerificationURI, m.deviceCode.UserCode)
-		output = markdownToString(m.width, prompt)
-
-	case Authenticated:
-		output = wrap(RenderOk()+" Authenticated successfully. Press any key to continue.", m.width-4, 2)
-	case ErrorAuthenticating:
-		output = wrap(RenderErr()+" Unable to authenticate. Please try again.", m.width-4, 2)
-	}
-
-	return containerStyle.Render(output)
-}
-
-type errMsg struct{ err error }
-type failedToAuthenticateErrorMsg struct{ err error }
-
-func openBrowserCmd(url string) tea.Cmd {
-	return func() tea.Msg {
-		err := browser.OpenURL(url)
-		if err != nil {
-			return displayAuthorizationInstructionsMsg{deviceCode: nil, err: err}
-		}
-		return WaitingForConfirmation
-	}
-}
-
-func awaitToken(ctx context.Context, config oauth2.Config, deviceCode *oauth2.DeviceAuthResponse) tea.Cmd {
-	return func() tea.Msg {
-		token, err := config.DeviceAccessToken(ctx, deviceCode)
-		if err != nil {
-			return failedToAuthenticateErrorMsg{err}
-		}
-
-		return token
-	}
-}
 
 // getChangeUuid returns the UUID of a change, as selected by --uuid or --change, or a state with the specified status and having --ticket-link
 func getChangeUuid(ctx context.Context, oi OvermindInstance, expectedStatus sdp.ChangeStatus, ticketLink string, errNotFound bool) (uuid.UUID, error) {
@@ -606,6 +493,261 @@ func login(ctx context.Context, cmd *cobra.Command, scopes []string, writer io.W
 	ctx, _ = context.WithTimeout(ctx, timeout) // nolint:govet // the context will not leak as the command will exit when it is done
 
 	return ctx, oi, token, nil
+}
+
+func ensureToken(ctx context.Context, oi OvermindInstance, requiredScopes []string) (context.Context, *oauth2.Token, error) {
+	var token *oauth2.Token
+	var err error
+
+	// get a token from the api key if present
+	if apiKey := viper.GetString("api-key"); apiKey != "" {
+		token, err = getAPIKeyToken(ctx, oi, apiKey, requiredScopes)
+	} else {
+		token, err = getOauthToken(ctx, oi, requiredScopes)
+	}
+	if err != nil {
+		return ctx, nil, fmt.Errorf("error getting token: %w", err)
+	}
+
+	// Check that we actually got the claims we asked for. If you don't have
+	// permission auth0 will just not assign those scopes rather than fail
+	ok, missing, err := HasScopesFlexible(token, requiredScopes)
+	if err != nil {
+		return ctx, nil, fmt.Errorf("error checking token scopes: %w", err)
+	}
+	if !ok {
+		return ctx, nil, fmt.Errorf("authenticated successfully, but you don't have the required permission: '%v'", missing)
+	}
+
+	// store the token for later use by sdp-go's auth client. Note that this
+	// loses access to the RefreshToken and could be done better by using an
+	// oauth2.TokenSource, but this would require more work on updating sdp-go
+	// that is currently not scheduled
+	ctx = context.WithValue(ctx, sdp.UserTokenContextKey{}, token.AccessToken)
+
+	return ctx, token, nil
+}
+
+// Gets a token from Oauth with the required scopes. This method will also cache
+// that token locally for use later, and will use the cached token if possible
+func getOauthToken(ctx context.Context, oi OvermindInstance, requiredScopes []string) (*oauth2.Token, error) {
+	var localScopes []string
+
+	// Check for a locally saved token in ~/.overmind
+	if home, err := os.UserHomeDir(); err == nil {
+		var localToken *oauth2.Token
+
+		localToken, localScopes, err = readLocalToken(home, requiredScopes)
+
+		if err != nil {
+			log.WithContext(ctx).Debugf("Error reading local token, ignoring: %v", err)
+		} else {
+			// If we already have the right scopes, return the token
+			return localToken, nil
+		}
+	}
+
+	// If we need to get a new token, request the required scopes on top of
+	// whatever ones the current local, valid token has so that we don't
+	// keep replacing it
+	requestScopes := append(requiredScopes, localScopes...)
+
+	// Authenticate using the oauth device authorization flow
+	config := oauth2.Config{
+		ClientID: oi.CLIClientID,
+		Endpoint: oauth2.Endpoint{
+			AuthURL:       fmt.Sprintf("https://%v/authorize", oi.Auth0Domain),
+			TokenURL:      fmt.Sprintf("https://%v/oauth/token", oi.Auth0Domain),
+			DeviceAuthURL: fmt.Sprintf("https://%v/oauth/device/code", oi.Auth0Domain),
+		},
+		Scopes: requestScopes,
+	}
+
+	deviceCode, err := config.DeviceAuth(ctx,
+		oauth2.SetAuthURLParam("audience", oi.Audience),
+		oauth2.AccessTypeOffline,
+	)
+	if err != nil {
+		return nil, fmt.Errorf("error getting device code: %w", err)
+	}
+
+	var token *oauth2.Token
+
+	statusParagraph := pterm.DefaultParagraph.WithMaxWidth(MAX_TERMINAL_WIDTH)
+	err = browser.OpenURL(deviceCode.VerificationURI)
+	if err != nil {
+		statusParagraph.Println(RenderErr() + " Unable to open browser: " + err.Error())
+	}
+	pterm.Print(
+		markdownToString(MAX_TERMINAL_WIDTH, fmt.Sprintf(
+			beginAuthMessage,
+			deviceCode.VerificationURI,
+			deviceCode.UserCode,
+		)))
+
+	token, err = config.DeviceAccessToken(ctx, deviceCode)
+	if err != nil {
+		statusParagraph.Println(RenderErr() + " Unable to authenticate. Please try again.")
+		log.WithContext(ctx).WithError(err).Error("Error getting device code")
+		os.Exit(1)
+	}
+	pterm.Println(RenderOk() + " Authenticated successfully. Press any key to continue.\n")
+	err = keyboard.Listen(func(keyInfo keys.Key) (stop bool, err error) {
+		key := keyInfo.Code
+		if key == keys.CtrlC {
+			statusParagraph.Println(RenderErr() + " Cancelled")
+			os.Exit(1)
+		}
+		log.WithField("key", key).Debug("Received keyboard input")
+		return true, nil
+	})
+	if err != nil {
+		statusParagraph.Println(RenderErr() + " Error reading keyboard input: " + err.Error())
+		os.Exit(1)
+	}
+
+	if token == nil {
+		statusParagraph.Println(RenderErr() + " Error running program: no token received")
+		os.Exit(1)
+	}
+
+	tok, err := josejwt.ParseSigned(token.AccessToken, []jose.SignatureAlgorithm{jose.RS256})
+	if err != nil {
+		statusParagraph.Println("Error running program: received invalid token:", err)
+		os.Exit(1)
+	}
+	out := josejwt.Claims{}
+	customClaims := sdp.CustomClaims{}
+	err = tok.UnsafeClaimsWithoutVerification(&out, &customClaims)
+	if err != nil {
+		statusParagraph.Println("Error running program: received unparsable token:", err)
+		os.Exit(1)
+	}
+
+	if cmdSpan != nil {
+		cmdSpan.SetAttributes(
+			attribute.Bool("ovm.cli.authenticated", true),
+			attribute.String("ovm.cli.accountName", customClaims.AccountName),
+			attribute.String("ovm.cli.userId", out.Subject),
+		)
+	}
+
+	// Save the token locally
+	if home, err := os.UserHomeDir(); err == nil {
+		// Create the directory if it doesn't exist
+		err = os.MkdirAll(filepath.Join(home, ".overmind"), 0700)
+		if err != nil {
+			log.WithContext(ctx).WithError(err).Error("Failed to create ~/.overmind directory")
+		}
+
+		// Write the token to a file
+		path := filepath.Join(home, ".overmind", "token.json")
+		file, err := os.Create(path)
+		if err != nil {
+			log.WithContext(ctx).WithError(err).Errorf("Failed to create token file at %v", path)
+		}
+
+		// Encode the token
+		err = json.NewEncoder(file).Encode(token)
+		if err != nil {
+			log.WithContext(ctx).WithError(err).Errorf("Failed to encode token file at %v", path)
+		}
+
+		log.WithContext(ctx).Debugf("Saved token to %v", path)
+	}
+
+	return token, nil
+}
+
+// Gets a token using an API key
+func getAPIKeyToken(ctx context.Context, oi OvermindInstance, apiKey string, requiredScopes []string) (*oauth2.Token, error) {
+	log.WithContext(ctx).Debug("using provided token for authentication")
+
+	var token *oauth2.Token
+
+	if !strings.HasPrefix(apiKey, "ovm_api_") {
+		return nil, errors.New("OVM_API_KEY does not match pattern 'ovm_api_*'")
+	}
+
+	// exchange api token for JWT
+	client := UnauthenticatedApiKeyClient(ctx, oi)
+	resp, err := client.ExchangeKeyForToken(ctx, &connect.Request[sdp.ExchangeKeyForTokenRequest]{
+		Msg: &sdp.ExchangeKeyForTokenRequest{
+			ApiKey: apiKey,
+		},
+	})
+	if err != nil {
+		return nil, fmt.Errorf("error authenticating the API token: %w", err)
+	}
+	log.WithContext(ctx).Debug("successfully got a token from the API key")
+
+	token = &oauth2.Token{
+		AccessToken: resp.Msg.GetAccessToken(),
+		TokenType:   "Bearer",
+	}
+
+	// Check that we actually got the claims we asked for. If you don't have
+	// permission auth0 will just not assign those scopes rather than fail
+	ok, missing, err := HasScopesFlexible(token, requiredScopes)
+	if err != nil {
+		return nil, fmt.Errorf("error checking token scopes: %w", err)
+	}
+	if !ok {
+		return nil, fmt.Errorf("authenticated successfully, but your API key is missing this permission: '%v'", missing)
+	}
+
+	return token, nil
+}
+
+func readLocalToken(homeDir string, requiredScopes []string) (*oauth2.Token, []string, error) {
+	// Read in the token JSON file
+	path := filepath.Join(homeDir, ".overmind", "token.json")
+
+	token := new(oauth2.Token)
+
+	// Check that the file exists
+	if _, err := os.Stat(path); err != nil {
+		return nil, nil, err
+	}
+
+	// Read the file
+	file, err := os.Open(path)
+	if err != nil {
+		return nil, nil, fmt.Errorf("error opening token file at %v: %w", path, err)
+	}
+
+	// Decode the file
+	err = json.NewDecoder(file).Decode(token)
+	if err != nil {
+		return nil, nil, fmt.Errorf("error decoding token file at %v: %w", path, err)
+	}
+
+	// Check to see if the token is still valid
+	if !token.Valid() {
+		return nil, nil, errors.New("token is no longer valid")
+	}
+
+	claims, err := extractClaims(token.AccessToken)
+	if err != nil {
+		return nil, nil, fmt.Errorf("error extracting claims from token: %w", err)
+	}
+	if claims.Scope == "" {
+		return nil, nil, errors.New("token does not have any scopes")
+	}
+
+	currentScopes := strings.Split(claims.Scope, " ")
+
+	// Check that we actually got the claims we asked for.
+	ok, missing, err := HasScopesFlexible(token, requiredScopes)
+	if err != nil {
+		return nil, currentScopes, fmt.Errorf("error checking token scopes: %w", err)
+	}
+	if !ok {
+		return nil, currentScopes, fmt.Errorf("local token is missing this permission: '%v'", missing)
+	}
+
+	log.Debugf("Using local token from %v", path)
+	return token, currentScopes, nil
 }
 
 func getAppUrl(frontend, app string) string {
